pyproject.toml
[project]
name = "portfolio-tracker-deveroskp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.2",
    "fastapi[standard]>=0.124.0",
    "pydantic-settings>=2.12.0",
    "pymysql>=1.1.0",
    "sqlalchemy>=2.0.0",
    "cryptography>=41.0.0",
]
